==> python/2_IMDBscraper_Step2.ipynb <==
import pandas as pd
import numpy as np
import re
import csv

from bs4 import BeautifulSoup
import requests
import time, os 
import random
from fake_useragent import UserAgent

pd.set_option('display.max_rows', None)
pd.set_option('display.max_columns', None)

ua = UserAgent()


movies = pd.read_csv('./movies_2021_2017.csv')

movies.shape

movies.tail()

movies = movies.drop(['Unnamed: 0'], axis=1) 
movies.head()

# A df with the awards link


movie_awards_link = movies[['title', 'movie_id', 'movie_imdb_link']]

movie_awards_link['awards_link'] = ''

movie_awards_link.head()

for i in movie_awards_link.index:
    awards_link_temp = movie_awards_link['movie_imdb_link'][i].split('/?ref_=adv_li_tt')[0]
    movie_awards_link['awards_link'][i] = awards_link_temp + '/awards/?ref_=tt_awd'
    

movie_awards_link.head()


# Scraping the awards information

def scrape_awards(df):
    
    movie_awards_dict = {}  #initializing an empty dictionary to fill with information

    for i in df.index:
        title = df['title'][i]
        movie_id = df['movie_id'][i]
        awards_link = df['awards_link'][i]
        
        user_agent = {'User-agent': ua.random} # generates a random user agent.
        response = requests.get(awards_link,headers = user_agent).text
        soup = BeautifulSoup(response, "html.parser")

        # awards and nominations:
        try:
            awards_total = soup.select('#main > div:nth-child(1) > div > div.header > div > div')[0].get_text()
        except:  
            awards_total = 0
        
        time.sleep(.5+2*random.random()) # random time between requests.
        
        movie_awards_dict[df.index[i]] = [title]+[movie_id]+[awards_link]+[awards_total]

    # define headers
    headers = ['title','movie_id','awards_link', 'awards_total']
    
    # turn into a dataframe    
    movie_awards_dataframe = pd.DataFrame(movie_awards_dict).T
    movie_awards_dataframe.columns = headers
    
    #save the dataframe
    #movie_awards_dataframe.to_csv('awards_2022.csv') # CHANGE THE YEAR ACC TO YOUR INTEREST
    
    return movie_awards_dataframe


awards_df = scrape_awards(movie_awards_link)

awards_df.head()

# Repeating everything for another dataframe

movies_2016_2015 = pd.read_csv('./movies_2016_2015.csv')

movies_2016_2015 = movies_2016_2015.drop(['Unnamed: 0'], axis=1) 
movies_2016_2015.tail(3)

movie_awards_link_2 = movies_2016_2015[['title', 'movie_id', 'movie_imdb_link']]

movie_awards_link_2['awards_link'] = ''

for i in movie_awards_link_2.index:
    awards_link_temp_2 = movie_awards_link_2['movie_imdb_link'][i].split('/?ref_=adv_li_tt')[0]
    movie_awards_link_2['awards_link'][i] = awards_link_temp_2 + '/awards/?ref_=tt_awd'
    

movie_awards_link_2.head()


awards_2016_2015_df = scrape_awards(movie_awards_link_2)

awards_2016_2015_df.head(3)

# Validation data (year 2022)

movies_2022 = pd.read_csv('./movies_2022.csv')

movies_2022.tail()

movies_2022 = movies_2022.reset_index()
movies_2022 = movies_2022.rename(columns = {'Unnamed: 0':'title'})


movies_2022.shape

movies_2022.head()

movies_2022 = movies_2022.drop(['index'], axis=1) 
movies_2022.head()

movie_awards_link_2022 = movies_2022[['title', 'movie_id', 'movie_imdb_link']]

movie_awards_link_2022['awards_link'] = ''

movie_awards_link_2022.head()

for i in movie_awards_link_2022.index:
    awards_link_temp_2022 = movie_awards_link_2022['movie_imdb_link'][i].split('/?ref_=adv_li_tt')[0]
    movie_awards_link_2022['awards_link'][i] = awards_link_temp_2022 + '/awards/?ref_=tt_awd'
    

movie_awards_link_2022.head()


awards_2022_df = scrape_awards(movie_awards_link_2022)

awards_2022_df.head()
