==> python/3_IMDBscraper_Step3.ipynb <==
import pandas as pd
import numpy as np
import re
import csv

from bs4 import BeautifulSoup
import requests
import time, os 
import random
from fake_useragent import UserAgent

pd.set_option('display.max_rows', None)
pd.set_option('display.max_columns', None)

ua = UserAgent()


movies = pd.read_csv('./movies_2021_2017.csv')

movies.shape

movies.tail()

movies = movies.drop(['Unnamed: 0'], axis=1) 
movies.head()

def scrape_movie_details(df):
    
    movies_details_dict = {}  #initializing an empty dictionary to fill with information
    
    for i in df.index:
        title = df['title'][i]
        movie_id = df['movie_id'][i]
        movie_imdb_link = df['movie_imdb_link'][i]

        user_agent = {'User-agent': ua.random} # generates a random user agent.
        response = requests.get(movie_imdb_link,headers = user_agent).text
        soup = BeautifulSoup(response, "html.parser")
    
        #top writer
        try:
            top_writer = soup.select('div.sc-fa02f843-2.dwQKsL > div > div > ul > li:nth-child(2) > div > ul > li:nth-child(1) > a')[0].get_text()
        except:
            top_writer = []
    
        # top listed three actors/actresses 

        #star #1
        try:
            top_star_1 = soup.select('div.sc-fa02f843-2.dwQKsL > div > div > ul > li:nth-child(3) > div > ul > li:nth-child(1) > a')[0].get_text()
        except:
            top_star_1 = []
        
        # star #2
        try:
            top_star_2 = soup.select('div.sc-fa02f843-2.dwQKsL > div > div > ul > li:nth-child(3) > div > ul > li:nth-child(2) > a')[0].get_text()
        except:
            top_star_2 = []
        
        # star #3
        try:
            top_star_3 = soup.select('div.sc-fa02f843-2.dwQKsL > div > div > ul > li:nth-child(3) > div > ul > li:nth-child(3) > a')[0].get_text()
        except:
            top_star_3 = []
        
        # number of user reviews
        try:
            num_user_reviews = soup.find_all(class_="label")[0].findPrevious().text
        except:
            num_user_reviews = 0
        
        # number of critic reviews
        try:
            num_critic_reviews = soup.find(text=re.compile('Critic reviews')).findPrevious().findPrevious().text
        except:
            num_critic_reviews = 0
        
        # release date
        try:
            release_date = soup.select('div.sc-f65f65be-0.ktSkVi > ul > li:nth-child(1) > div > ul > li > a')[0].get_text().split('(')[0].strip()
        except:
            release_date = []
        
        # release month
        try:
            release_month = soup.select('div.sc-f65f65be-0.ktSkVi > ul > li:nth-child(1) > div > ul > li > a')[0].get_text().split('(')[0].strip().split(' ')[0].strip()
        except: 
            release_month = []
        
        # country of origin
        try:
            country_of_origin = soup.select('div.sc-f65f65be-0.ktSkVi > ul > li:nth-child(2) > div > ul > li > a')[0].get_text()
        except:
            country_of_origin = []
        
        # top production company:
        try:
            top_production_company = soup.select('div.sc-f65f65be-0.ktSkVi > ul > li:nth-child(7) > div > ul > li:nth-child(1) > a')[0].get_text()
        except:
            top_production_company = []
        
        # Budget in USD: Production budget refers to the cost to make the movie and it does not include marketing or other expenditures.
        try:
            budget_in_usd = soup.select('div.sc-f65f65be-0.ktSkVi > ul > li:nth-child(1) > div > ul > li > span')[0].get_text().split('(')[0].strip().replace('$',"").replace(',','')
        except:
            budget_in_usd = []
        
        # Opening weekend US & Canada (in USD): Weekend box office charts show gross receipts for a given weekend, which is Friday through Sunday unless otherwise noted.
        try:
            opening_weekend_us_can_in_usd = soup.select('div.sc-f65f65be-0.ktSkVi > ul > li:nth-child(3) > div > ul > li:nth-child(1) > span')[0].get_text().replace('$',"").replace(',','')
        except: 
            opening_weekend_us_can_in_usd = []
        
        # Gross US & Canada in USD (Gross refers to gross earnings in U.S. dollars.)
        try:
            gross_us_can_in_usd = soup.select('div.sc-f65f65be-0.ktSkVi > ul > li:nth-child(2) > div > ul > li > span')[0].get_text().replace('$',"").replace(',','')
        except:
            gross_us_can_in_usd = []
        
        # Gross worldwide in USD
        try: 
            gross_worldwide_in_usd = soup.select('div.sc-f65f65be-0.ktSkVi > ul > li:nth-child(4) > div > ul > li > span')[0].get_text().replace('$',"").replace(',','')
        except:
            gross_worldwide_in_usd = []
        
        movies_details_dict[df.index[i]] = [title]+[movie_id]+[movie_imdb_link]+[top_writer]+[top_star_1]+[top_star_2]+\
        [top_star_3]+[num_user_reviews]+[num_critic_reviews]+[release_date]+[release_month]+[country_of_origin]+\
        [top_production_company]+[budget_in_usd]+[opening_weekend_us_can_in_usd]+[gross_us_can_in_usd]+\
        [gross_worldwide_in_usd]
    
        time.sleep(.5+2*random.random()) # random time between requests.

    # define headers
    headers = ['title','movie_id','movie_imdb_link', 'top_writer','top_star_1','top_star_2','top_star_3',
                  'num_user_reviews', 'num_critic_reviews', 'release_date','release_month','country_of_origin',
                  'top_production_company','budget_in_usd', 'opening_weekend_us_can_in_usd','gross_us_can_in_usd',
                  'gross_worldwide_in_usd' ]
    
    # turn into a dataframe    
    movie_details_dataframe = pd.DataFrame(movies_details_dict).T
    movie_details_dataframe.columns = headers
    
    #save the dataframe
    #movie_details_dataframe.to_csv('movie_details_2022.csv') # CHANGE THE YEAR ACC TO YOUR INTEREST
    
    return movie_details_dataframe


movie_details = scrap_movie_details(movies)


movie_details.shape

movie_details.head(10)

# Repeating everything for another dataframe

movies_2016_2015 = pd.read_csv('./movies_2016_2015.csv')

movies_2016_2015 = movies_2016_2015.drop(['Unnamed: 0'], axis=1) 
movies_2016_2015.head()

movie_details_2016_2015 = scrape_movie_details(movies_2016_2015)

movie_details_2016_2015.head(3)

# Validation data (year 2022)

movies_2022 = pd.read_csv('./movies_2022.csv')

movies_2022.head()

movies_2022 = movies_2022.reset_index(drop=True)
movies_2022 = movies_2022.rename(columns = {'Unnamed: 0':'title'})
movies_2022.head()

movie_details_2022 = scrape_movie_details(movies_2022)

movie_details_2022.head(3)

movie_details_2022.shape
