==> python/4_Academy_awards_df_(adding_2021).ipynb <==
import pandas as pd
import numpy as np
import re
import csv

from bs4 import BeautifulSoup
import requests
import time, os 
import random
from fake_useragent import UserAgent

pd.set_option('display.max_rows', None)
pd.set_option('display.max_columns', None)

ua = UserAgent()

oscars = pd.read_csv('./oscarnominees.csv', encoding="ISO-8859-1")

oscars.head()

oscars.Award.value_counts()

oscars_2021_dict = {'Title': ['Belfast', 'CODA', "Don't Look Up" , 'Drive My Car', 'Dune', 'King Richard', 'Licorice Pizza', 'Nightmare Alley', 'The Power of the Dog', 'West Side Story',
                             'Jessica Chastain','Olivia Colman', 'Penélope Cruz', 'Nicole Kidman','Kristen Stewart',
                             'Javier Bardem', 'Benedict Cumberbatch', 'Andrew Garfield', 'Will Smith', 'Denzel Washington',
                             'Ciarán Hinds', 'Troy Kotsur', 'Jesse Plemons', 'J.K. Simmons', 'Kodi Smit-McPhee',
                             'Jessie Buckley', 'Ariana DeBose', 'Judi Dench', 'Kirsten Dunst', 'Aunjanue Ellis'],
                    'Winner': ['N', 'Y', "N" , 'N', 'N', 'N', 'N', 'N', 'N', 'N',
                              'Y', "N" , 'N', 'N', 'N',
                              'N', "N" , 'N', 'Y', 'N',
                              'N', "Y" , 'N', 'N', 'N',
                              'N', "Y" , 'N', 'N', 'N'],
                    'Award': ['Picture', 'Picture', "Picture" , 'Picture', 'Picture', 'Picture', 'Picture', 'Picture', 'Picture', 'Picture',
                             'Actress','Actress', 'Actress', 'Actress', 'Actress',
                             'Actor', 'Actor', 'Actor', 'Actor', 'Actor',
                             'Supporting Actor', 'Supporting Actor','Supporting Actor','Supporting Actor','Supporting Actor',
                             'Supporting Actress', 'Supporting Actress', 'Supporting Actress', 'Supporting Actress', 'Supporting Actress'],
                    'Year': ['2021', '2021', "2021" , '2021', '2021', '2021', '2021', '2021', '2021', '2021',
                              '2021', "2021" , '2021', '2021', '2021',
                              '2021', "2021" , '2021', '2021', '2021',
                              '2021', "2021" , '2021', '2021', '2021',
                              '2021', "2021" , '2021', '2021', '2021']}



oscars_2021_df = pd.DataFrame.from_dict(oscars_2021_dict)
  


oscars_2021_df

oscars_1928_2021_major_categories = pd.concat([oscars_2021_df,oscars],axis = 0) 

oscars_1928_2021_major_categories = oscars_1928_2021_major_categories.reset_index(drop=True)


#oscars_1928_2021_major_categories

oscars_1928_2021_major_categories.to_csv('oscars_1928_2021_major_categories.csv')


#Source for 2021: https://awardsdatabase.oscars.org/search/results
