==> python/6_DataPrep_ValidationDataset.ipynb <==
#main
import pandas as pd
import numpy as np
import math
import sklearn
import re

#graphic
import matplotlib
import matplotlib.pyplot as plt
import seaborn as sns 
%matplotlib inline

pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)

movie_overview_2022 = pd.read_csv('./movies_2022.csv', index_col='Unnamed: 0')
movie_detail_2022 = pd.read_csv('./movie_details_2022.csv', index_col='Unnamed: 0')
movie_award_2022 = pd.read_csv('./awards_2022.csv', index_col='Unnamed: 0')
top_1000_movies = pd.read_csv('./imdb_top_1000.csv')


movie_overview_2022 = movie_overview_2022.reset_index(drop=True)

movie_detail_2022.drop(['title','movie_id','movie_imdb_link' ], axis=1, inplace=True) # drop duplicate columns

movie_award_2022.drop(['title','movie_id'], axis=1, inplace=True) # drop duplicate columns

movies_df = pd.concat([movie_overview_2022, movie_award_2022, movie_detail_2022 ],axis = 1)


movies_df.shape

def separete_awards(df):
    df['total_award_nominations'] = '' #creating empty columns
    df['total_award_wins'] = ''

    for i in df.index:
        if df['awards_total'][i] != '0':
            pattern= '\d+'
            numbers = re.findall(pattern,df['awards_total'][i])
            df['total_award_wins'][i] = numbers[0]      
            df['total_award_nominations'][i] = numbers[1]         
      
        else:
            df['total_award_wins'][i] = df['awards_total'][i]
            df['total_award_nominations'][i] = df['awards_total'][i]
    
    return df

movies_df_v1 = separete_awards(movies_df)

def separete_genre(df):
    df['primary_genre'] = '' #creating empty columns
    df['secondary_genre'] = ''
    
    pattern = "[\w']+"
    
    for i in df.index:
        if df['genre'][i] != '[]':
            g = re.findall(pattern,df['genre'][i])
            
            if len(g)>1:
                df['primary_genre'][i] = g[0]      
                df['secondary_genre'][i] = g[1]
            else:
                df['primary_genre'][i] = df['genre'][i]
                df['secondary_genre'][i] = 'Other'
        else:
            df['primary_genre'][i] = ''
            df['secondary_genre'][i] = ''
    
    return df
    

movies_df_v2 = separete_genre(movies_df_v1)

movies_df_v2['certificate']= movies_df_v2['certificate'].replace({
    'R':'Adult',
    'PG-13':'+13/14', 
    'TV-13':'+13/14',
    'TV-MA':'Adult',
    'PG': 'Parental Guidance',
    'TV-14': '+13/14',
    'Unrated': 'Not Rated',
    'Approved': 'Not Rated',
    'TV-PG': 'Parental Guidance',
    'TV-G': 'Kids',
    'G': 'Suitable for all',
    'TV-Y7': 'Kids',
    '18': 'Adult',
    'TV-Y': 'Kids',
    'TV-Y7-FV': 'Kids',
    'T':'Adult',
    'M': 'Adult'
})

def drop_k(df):
    
    pattern = "K"
    
    for i in df.index:
        if df['num_user_reviews'][i] != '[]':
            review = re.findall(pattern,df['num_user_reviews'][i]) 
            if len(review)>0:
                a = df['num_user_reviews'][i].replace('K','')    
                df['num_user_reviews'][i] = float(a)*1000
                
            else:
                df['num_user_reviews'][i] = df['num_user_reviews'][i]
        else:
            df['num_user_reviews'][i] = 0
    
    df['num_user_reviews'] =  pd.to_numeric(df['num_user_reviews'], errors='coerce')
    
    return df
    

movies_df_v2 = drop_k(movies_df_v2)

movies_df_v2['num_critic_reviews'] =  pd.to_numeric(movies_df_v2['num_critic_reviews'], errors='coerce')

movies_df_v2['release_date'].value_counts(dropna=False)

movies_df_v2['release_date']= movies_df_v2['release_date'].replace({
    '2022':'',
    '2023':''
})

movies_df_v2['release_date']=pd.to_datetime(movies_df_v2['release_date'], errors='coerce')

movies_df_v2['release_date'].value_counts(dropna=False)

import datetime as dt

movies_df_v2['release_weekday'] = movies_df_v2['release_date'].dt.isocalendar().day # weekday 5 is a friday


movies_df_v2 = movies_df_v2.drop(['release_month'], axis=1) #dropping the original month column and 
#creating one from datetime. The original release_month column has more nonsensical values


movies_df_v2['release_month'] = pd.DatetimeIndex(movies_df_v2['release_date']).month


movies_df_v2['release_month'].value_counts(dropna=False)

movies_df_v2['release_year'].value_counts(dropna=False)

movies_df_v2['country_of_origin'].value_counts(dropna=False)

movies_df_v2['metascore'].value_counts(dropna=False)

movies_df_v2['metascore']= movies_df_v2['metascore'].replace({
    '[]':'Not scored'    
})


def remove_currency(row): 
    if type(row) != float:
        row = row.replace('$', '')
        if '€' in row:
            row = row.replace('€', '').strip() # no cunversion needed as of date
        if 'CA' in row:
            row = row.replace('CA', '').strip()
            row = str(int(row)*0.73) 
        if '₹' in row:
            row = row.replace('₹', '').strip()
            row = str(int(row)*0.012) 
        if '£' in row:
            row = row.replace('£', '').strip()
            row = str(int(row)*1.14)
        if 'CN¥' in row:
            row = row.replace('CN¥', '').strip()
            row = str(int(row)*0.0067)
        if 'RUR' in row:
            row = row.replace('RUR', '').strip()
            row = str(int(row)*0.016)
             
    return row

movies_df_v2['budget_in_usd'] = movies_df_v2['budget_in_usd'].apply(remove_currency)

def remove_nonsensical(row): 
    if type(row) != float:
        if ':' in row:
            row = np.nan
            
    return row

movies_df_v2['opening_weekend_us_can_in_usd'] = movies_df_v2['opening_weekend_us_can_in_usd'].apply(remove_nonsensical)

movies_df_v2['gross_us_can_in_usd'] = movies_df_v2['gross_us_can_in_usd'].apply(remove_nonsensical)

movies_df_v2['gross_worldwide_in_usd'] = movies_df_v2['gross_worldwide_in_usd'].apply(remove_nonsensical)


movies_df_v2.isna().sum()

movies_df_v2.replace('',np.nan, inplace=True)
movies_df_v2.replace('[]',np.nan, inplace=True)

movies_df_v2.isna().sum()

cols_numeric = ['runtime_in_mins','num_user_reviews', 'metascore','budget_in_usd', 'opening_weekend_us_can_in_usd',
            'gross_us_can_in_usd', 'gross_worldwide_in_usd','total_award_nominations',
            'total_award_wins']

movies_df_v2[cols_numeric] = movies_df_v2[cols_numeric].apply(pd.to_numeric, errors='coerce', axis=1)


cols_object = ['release_year','release_month', 'release_weekday']

movies_df_v2[cols_object] = movies_df_v2[cols_object].astype(object)

movies_df_v2.isna().sum()

# Feature engineering

## Oscars proved to be useless for the model.

## Director

top_1000_movies_copy = top_1000_movies.copy()
best_directors =top_1000_movies_copy.pivot_table(index= ['Director'],  aggfunc = ['count'])
best_directors.columns = best_directors.columns.droplevel(0)
best_directors = best_directors.reset_index().rename_axis(None, axis=1)


best_directors = best_directors.drop(['Genre', 'Gross', 'IMDB_Rating',
       'Meta_score', 'No_of_Votes', 'Overview', 'Poster_Link', 'Released_Year',
       'Runtime', 'Series_Title', 'Star1', 'Star2', 'Star3', 'Star4'], axis=1)

best_directors = best_directors.rename(columns={'Director':'top_director',
                                    'Certificate':'is_among_best_director'})

best_directors['is_among_best_director'] = 'Y'

best_directors.shape


movies_df_v2.shape

movies_df_v10 = pd.merge(movies_df_v2, best_directors, how='left', on='top_director')

movies_df_v10.tail(10)

movies_df_v10_bud = movies_df_v10[~(
    movies_df_v10['budget_in_usd'] < 10000)]

movies_df_v10_gross_uscan = movies_df_v10_bud[~(
    movies_df_v10['gross_us_can_in_usd'] < 2000)]

movies_df_v10_gross_uscan.isna().sum()

# I will continue with a filtered data where the financial columns are not null.

mask1 = movies_df_v10_gross_uscan['gross_worldwide_in_usd'].notnull()
mask2 = movies_df_v10_gross_uscan['opening_weekend_us_can_in_usd'].notnull()
mask3 = movies_df_v10_gross_uscan['budget_in_usd'].notnull()

movies_filtered = movies_df_v10_gross_uscan[mask1 & mask2 & mask3].copy()

movies_filtered.shape

movies_filtered.isna().sum()

movies_filtered['release_weekday'].value_counts()

movies_filtered['release_weekday'] = movies_filtered['release_weekday'].fillna(5).astype(object)


movies_filtered['release_month'].value_counts()

movies_filtered['release_month'] = movies_filtered['release_month'].fillna(4.0).astype(object)
movies_filtered['release_month'].value_counts()


def group_prod_company(df):

    major_prod_company = ['Universal Pictures', 'Columbia Pictures', 'Warner Bros.', 'Paramount Pictures']
    df['top_production_company_grouped'] = '' #creating an empty column
    
    for i in df.index:
        if df['top_production_company'][i] not in major_prod_company:
            df['top_production_company_grouped'][i] = 'Other'      
         
        else:
            df['top_production_company_grouped'][i] = df['top_production_company'][i]
  
    return df



movies_filtered_v1 = group_prod_company(movies_filtered)

movies_filtered_v1['top_production_company_grouped'].value_counts(dropna=False)

movies_filtered_v1['top_production_company']= movies_filtered_v1['top_production_company'].fillna('Other')

movies_filtered_v1['top_production_company'].isna().sum()

def group_country(df):

    df['country_of_origin_grouped'] = '' #creating an empty column
    
    for i in df.index:
        if df['country_of_origin'][i] != 'United States':
            df['country_of_origin_grouped'][i] = 'Other'      
         
        else:
            df['country_of_origin_grouped'][i] = df['country_of_origin'][i]
  
    return df



movies_filtered_v2 = group_country(movies_filtered_v1)

movies_filtered_v2['primary_genre']= movies_filtered_v2['primary_genre'].replace({'Fantasy': 'Other',
        'Thriller': 'Other','Romance': 'Other', 'Music': 'Other', 'Mystery': 'Other', 'Musical': 'Other',
        'Sci':'Other' })

movies_filtered_v2['primary_genre'].value_counts(dropna=False)


movies_filtered_v2['secondary_genre']= movies_filtered_v2['secondary_genre'].replace({'Mystery': 'Other',
        'Fantasy': 'Other','Family': 'Other', 'Biography': 'Other', 'Action': 'Other', 'Sci': 'Other',
        'History': 'Other','Music': 'Other', 'Sport': 'Other', 'War': 'Other', 'Musical': 'Other', 
        'Western': 'Other', 'Fi': 'Other'})

movies_filtered_v2['secondary_genre'].value_counts(dropna=False)


movies_filtered_v2['certificate']=movies_filtered_v2['certificate'].fillna('Not Rated')
movies_filtered_v2['certificate'].value_counts(dropna=False)

movies_filtered_v2['num_critic_reviews']=movies_filtered_v2['num_critic_reviews'].fillna(0)

movies_filtered_v2.isna().sum().sum()

movies_filtered_v2['num_critic_reviews']=movies_filtered_v2['num_critic_reviews'].fillna(0)

movies_filtered_v2.columns

movies_filtered_v2['is_among_best_director']= movies_filtered_v2['is_among_best_director'].fillna('N')

movies_filtered_v2.isna().sum().sum()

movies_filtered_v2.to_csv('movies_2022_treated.csv')
